небольшая уборка кода
