+++ src/view/trip-events-add-point-view.js
@@ -3,7 +3,8 @@ import dayjs from 'dayjs';
 import { markUpDestinationPhotos } from '../template/pictures.js';
 import { markUpOfferSelectores } from '../template/offers-selector.js';
 
-const EVENT_TYPES = ['Taxi', 'Bus', 'Train', 'Ship', 'Drive', 'Flight', 'Check-in', 'Sightseeing', 'Restaurant'];
+import {EVENT_TYPES,defaultPoint,defaultDestination} from '../mock/const.js';
+
 
 const createEventTypeTemplate = (type, pointType, id) => `
   <div class="event__type-item">
@@ -11,28 +12,6 @@ const createEventTypeTemplate = (type, pointType, id) => `
     <label class="event__type-label  event__type-label--${type.toLowerCase()}" for="event-type-${type.toLowerCase()}-${id}">${type}</label>
   </div>
 `;
-
-const defaultPoint = {
-  id: '',
-  basePrice: 0,
-  dateFrom: '2024-05-14T07:27:07.173Z',
-  dateTo: '2024-05-15T08:12:07.173Z',
-  isFavorite: false,
-  offers: [],
-  type: EVENT_TYPES[0],
-};
-
-const defaultDestination = {
-  id: 'baedd0da-d74e-45ef-ad35-ee37057c3242',
-  description: 'Kioto - with an embankment of a mighty river as a centre of attraction',
-  name: 'Kioto',
-  pictures: [
-    {
-      src: 'https://23.objects.htmlacademy.pro/static/destinations/17.jpg',
-      description: 'Kioto with crowded streets',
-    },
-  ],
-};
 
 function createTripEventsAddPointElements() {
   const { type, dateFrom, dateTo, basePrice, id } = defaultPoint;
